3 варианты обработки указателя: просто клик, клик+исчезновение, клик+перемещение+исчезновение

+++ domain/src/main/java/com/example/domain/model/CharData.kt
@@ -2,6 +2,6 @@ package com.example.domain.model
 
 data class CharData(
     val charByte: Byte,
-    val colorByte: Byte,
-    val backgroundByte: Byte
+    val colorByte: Int,
+    val backgroundByte: Int
 )
